2026年4月21日 16:53

+++ 东方博宜2026.3月B组/C.cpp
@@ -21,7 +21,7 @@ void solve(){
         for(auto j:mp[i]){
             if(j>ans%s.size()){
                 f=1;
-                ans=(ll)(ans/s.size())+j;
+                ans=ans/s.size()+j;
                 cerr<<j<<" ";
                 break;
             }
